+++ app/Containers/User/Data/Seeders/UserPermissionsSeeder_1.php
@@ -2,9 +2,8 @@
 
 namespace App\Containers\User\Data\Seeders;
 
-use App\Containers\Authorization\Tasks\CreatePermissionTask;
+use Apiato\Core\Foundation\Facades\Apiato;
 use App\Ship\Parents\Seeders\Seeder;
-use Illuminate\Support\Facades\App;
 
 /**
  * Class UserPermissionsSeeder_1
@@ -21,12 +20,11 @@ class UserPermissionsSeeder_1 extends Seeder
     public function run()
     {
         // Default Permissions ----------------------------------------------------------
-
-        App::make(CreatePermissionTask::class)->run('search-users', 'Find a User in the DB.');
-        App::make(CreatePermissionTask::class)->run('list-users', 'Get All Users.');
-        App::make(CreatePermissionTask::class)->run('update-users', 'Update a User.');
-        App::make(CreatePermissionTask::class)->run('delete-users', 'Delete a User.');
-        App::make(CreatePermissionTask::class)->run('refresh-users', 'Refresh User data.');
+        Apiato::call('Authorization@CreatePermissionTask', ['search-users', 'Find a User in the DB.']);
+        Apiato::call('Authorization@CreatePermissionTask', ['list-users', 'Get All Users.']);
+        Apiato::call('Authorization@CreatePermissionTask', ['update-users', 'Update a User.']);
+        Apiato::call('Authorization@CreatePermissionTask', ['delete-users', 'Delete a User.']);
+        Apiato::call('Authorization@CreatePermissionTask', ['refresh-users', 'Refresh User data.']);
 
         // ...