+++ bootstrap/app.php
@@ -2,6 +2,19 @@
 
 /*
 |--------------------------------------------------------------------------
+| Set PHPs serialization precision 
+|--------------------------------------------------------------------------
+| Make sure that everyone who has the right php version is using the more 
+| precise serialization_precision
+| refer to: https://github.com/apiato/apiato/issues/257 for an explanation
+| of the issue.
+*/
+if (version_compare(phpversion(), '7.1', '>=')) {
+    ini_set( 'serialize_precision', -1 );
+}
+
+/*
+|--------------------------------------------------------------------------
 | Create The Application
 |--------------------------------------------------------------------------
 |